api/requirements.txt
google-generativeai>=0.3.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0
gspread>=5.12.0
oauth2client>=4.1.3
PyPDF2>=3.0.0
